Refactor CSV processor for keeping single responsibility principle
